+++ tests/test_geometric_search.py
@@ -34,10 +34,10 @@ import random
 import numpy as np
 import pytest
 
-from meshpy.core.element_beam import Beam3rHerm2Line3
 from meshpy.core.material import MaterialReissner
 from meshpy.core.mesh import Mesh
 from meshpy.core.rotation import Rotation
+from meshpy.four_c.element_beam import Beam3rHerm2Line3
 from meshpy.geometric_search.find_close_points import (
     FindClosePointAlgorithm,
     find_close_points,
